KAN-282 refactor: 상품리뷰 평균별점 로직 리팩토링

+++ src/main/java/com/yeonieum/productservice/domain/review/dto/ProductReviewResponse.java
@@ -14,6 +14,6 @@ public class ProductReviewResponse {
         private LocalDateTime createDate;
         private String reviewContent;
         private String reviewImage;
-        private int reviewScore;
+        private double reviewScore;
     }
 }
